HURRAY
Just lost a shitrack of work but I retyped it all hurray.
There aren't any comments in it because those take a while and I have to
run to build.

Dear TA,

Here's a (sort of) changelog:
Added MainActivity2 which just says "Settings", though that can be
easily changed by choosing a different layout.  (Didn't know what kind
of things we want to go in here)
Made it so the settings action bar item is in the overflow and opens
MainActivity2 via openSettings() method using intents

Adjusted meta-data to recognize MainActivity as the parent activity
across SDK versions.  Also added the "Up Button" functionality

Probably other things but I have to run to build call bye